+++ Demo/VComponentsDemo/Scenes/Demos/Modals/VBottomSheetDemoView.swift
@@ -145,7 +145,7 @@ struct VBottomSheetDemoView: View {
         VCheckBox(
             model: {
                 var model: VCheckBoxModel = .init()
-                model.layout.titleLabelLineLimit = 1
+                model.layout.titleLineLimit = 1
                 return model
             }(),
             isOn: .init(