+++ Demo/VComponentsDemo/Scenes/Demos/Modals/VModalDemoView.swift
@@ -92,7 +92,7 @@ struct VModalDemoView: View {
         VCheckBox(
             model: {
                 var model: VCheckBoxModel = .init()
-                model.layout.titleLabelLineLimit = 1
+                model.layout.titleLineLimit = 1
                 return model
             }(),
             isOn: .init(