+++ Sources/VComponents/Components/State Pickers/Check Box/VCheckBox.swift
@@ -127,7 +127,7 @@ public struct VCheckBox<Label>: View where Label: View {
 
                     VBaseButton(gesture: gestureHandler, label: {
                         VText(
-                            type: .multiLine(alignment: .leading, lineLimit: model.layout.titleLabelLineLimit),
+                            type: .multiLine(alignment: .leading, lineLimit: model.layout.titleLineLimit),
                             color: model.colors.title.for(internalState),
                             font: model.fonts.title,
                             title: title